rental_service/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
requests==2.32.3
gunicorn==23.0.0
whitenoise==6.8.2
python-decouple==3.8
